Use price price_1R7fOMLWue9IBlPSrEgHqsMg instead of price_1QMtlZLWue9IBlPSnboRsJbz (#1586)
The usage-based price item that we were using on Stripe for Tuist Pro,
`price_1QMtlZLWue9IBlPSnboRsJbz`, was wrongly configured, which led to
not applying the "200 hits for free". As shown in the screenshot below,
it needs to be configured as "graduated" and not "volume", which is the
one that we were using in `price_1QMtlZLWue9IBlPSnboRsJbz`.

<img width="587" alt="image"
src="https://github.com/user-attachments/assets/9b55bcfe-59a8-4470-beca-8f7aac2566c9"
/>

> [!NOTE]
> Since we didn't have that many subscriptions on that plan, I manually
fixed them through Stripe's UI.