Don't create Stripe customers on account creation (#1598)
We are currently creating Stripes on account creation, something that's
not really necessary until the account plans to upgrade from the air to
the pro plan.

This PR removes the customer creation on account creation (the column
was already optional), and updates logic throughout the codebase that
assumes that `account.customer_id` is presents.