Handle Swift packages that have directories at the root that end with `Package.swift` (#1615)
Our service came across [a
package](https://github.com/pytorch/executorch/tree/main/.Package.swift)
that decided that having a directory called `.Package.swift` was a good
idea :). Our logic assumes that a file path whose base component ends in
`Package.swift` is a manifest and wrongly treats it as a file. I'm
extending the logic to check if the path represents a file, and skip
otherwise.