Fix unknown hash in the run detail (#1660)
In the production, the hashes for binary cache are always `Unknown`
whereas hashes for selective testing do appear, such as here:
https://tuist.dev/tuist/tuist/runs/989889

The problem was that the deprecated method of listing out targets was
used as it declared first, taking precedence. Changing the order fixes
the issue.

When writing the test, I also noticed a bunch of `has_element?` calls
didn't have the `assert`, so they were not really testing anything 😅 So,
there are a bunch of fly-by fixes related to that.