Remove references to the old 'Tuist Cloud' (#1658)
I noticed we have some references in recent articles to "Tuist Cloud", a
term we are no longer using, so I went through those and updated them.

I must admit that finding an alternative term to "Tuist Cloud" is
tricky. I debated between "Server", which I don't like, or just
mentioning that some of the features are "Paid". Since the future of
Tuist is providing most of its value through the server, I expect this
distinction to be less critical since people will associate Tuist with
being a service that you plug into your project.