Read target project from the connection assigns (#1616)
I'm doing a bit of a refactoring to move us away from the pattern of
using `EnsureProjectPresencePlug`'s interface to get the project that a
request is targeting. I propose that we use the naming convention
`:target_project` to locate the project in the assigns (akin to
`:current_`) for the authenticated subject.

In follow-up PRs I'll do something similar with the authenticated
subject, and then decoupling the loading of a resource from checking its
presence.