Fix the updating of `remote_cache_hits_count` on Stripe (#1667)
[AppSignal
issue](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/724)

By making `customer_id` an optional attribute in `accounts`, we
introduced a regression in the job that goes through all the accounts,
updating the `remote_cache_hits_count` meter in their Stripe accounts.
This means that we have some data loss for some accounts, but it's not
the end of the world.

This PR not only fixes that, but also introduces other improvements:

- I've adjusted the implementation to paginate using Flop instead of
Stream, which runs inside a transaction, and we blocked the database
until all the updates were complete. This isn't good. But we were too
junior with Elixir and Ecto :)
- I schedule a job per account update, using jobs as an error boundary,
and giving us the flexibility to define individual retry rules.