feat(ui): create project flow (#1648)
This PR creates the "create a new project, and optionally a new
organization" flow after confirming a user's account.
To accomplish that it
- adds a new select hook and component
- adds two new routes
- refactors some existing domain code to allow for better error handling
in liveviews, given that this functionality was previously only
available via API

I'll push the follow up home page empty state in a separate PR, so
currently nothing happens on project creation. That's intended 👀



https://github.com/user-attachments/assets/1249a755-8e9f-477f-bed9-ac61bfd03f02