Revert the reverts (#1675)
I had to hotfix-rollout two reverts
([this](https://github.com/tuist/server/pull/1673) and
[this](https://github.com/tuist/server/pull/1674)) in the aim of fixing
production errors.

It turns out they happened because I enabled the new design for
everyone, and that surfaced that we had some connection redirects that
that were missing a halt. Therefore, Phoenix continued routing them,
leading to a redirect on a redirected `conn`.

This PR reverts the reverts and adds the missing `halt(conn)`