Add cache to the steps that build Docker images (#1678)
I followed [this
tutorial](https://www.blacksmith.sh/blog/cache-is-king-a-guide-for-docker-layer-caching-in-github-actions)
to enable caching in the steps that build with `docker` using GitHub
Actions. Note that the cache is evicted automatically by GitHub when it
goes above 10GB, but that should be fine, since we don't really need old
cache layers.