First step towards consolidating resource loading in a single plug (#1695)
The current logic for loading `selected_` resources pointed by URL or
body params is scattered across several plugs and controller methods,
making the logic a bit tricky to reason about and optimize. Moreover, we
have those `Ensure...` plugs through which we load and fetch resources,
which feels a bit unnecessary in the fetch part because we could instead
resort to a naming convention to directly read the resources from the
assigns (e.g., `conn.assigns[:selected_account]`).

This PR introduces a loader plug over which I plan to move all the
resource loading logic. I propose that we make this plug reusable across
API (e.g., JSON) and web endpoints.

> [!NOTE]
> In the long term, every request should at maximum do 3 SQL queries,
one to fetch the `current_` entity, another one for `selected_`, and
then the query to fetch the requested resource.