Strenthen the password policy and ensure it runs on password update (#1697)
Fixes https://github.com/tuist/server/issues/1652

The most recent penetration test uncovered that our password policy was
weak (we only checked the size) and that we didn't run it when changing
the password. This PR solves it by using
[zxcvbn-elixir](https://github.com/techgaun/zxcvbn-elixir), a port of a
tool written by Dropbox to check the strength of a password. I included
it in the user changesets that we run when creating users and updating
their passwords.