Configure the DB to keep the connection alive (#1745)
Supabase mentioned that with our DB size, the number of max connections
should be 160:

> The Large instance also allows 160 max direct DB connections and 800
max pooler clients. The average number of client connections for the
previous 30 days is around ~32. Here is the client connections graph
from [Grafana](https://supabase.com/docs/guides/telemetry/metrics):

The dashboard wrongly mentioned that it's 20 (I'll report it to them),
so I'm bumping it.

> [!NOTE]
> I doubt the issue that we are seeing is related to the exhaustion of
the client pool, but since we have resources to handle more, there's no
need to have them in a queue. It's DB capacity that we are not taking
advantage of.