chore: add styler as formatter (#1633)
Adds [Adobe's Styler](https://github.com/adobe/elixir-styler) to the
project.
It's basically running `lint-fix` in the CLI repo, just for Elixir. A
fix mode for a ton of Credo rules.

This will need a bunch of review (and mangled some comments in the
`config` dir that I will fix) but would like to hear about what you
think about adding it in general :)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>