Integrate the remote instances (production, canary, and staging) with Upstash-managed Redis instances (#1749)
The implementation of gradual backoffs when rate-limiting the auth
endpoints will require persisting the backoff state associated with the
IP address in a storage solution that persists across deployments. Since
this is not something that we want to leak into the DB, I set up 3 Redis
instances in Fly for production, canary, and staging, and I'm
configuring the `:redix` dependency in this project to connect to those
instances.