Not use `Repo.transaction` and `Repo.stream` when updating accounts' current month usage (#1752)
### Context
We have a background job that updates the current month's usage such
that we can present this information in the dashboard without having to
run an expensive query. The current implementation uses a single job
using a stream inside a transaction. This is not ideal since one failure
would cause the job to exit leading to the remaining accounts not
getting updated.

### Solution
I separated the worker into two, one that traverses all the accounts and
schedules the update, and another whose scope is updating the usage of
each account. That eliminates the possibility of one failure causing
everything else to stop. Moreover, I changed the approach from using
streams to paginated queries using flops.