Supersede `EnsureAccountPresencePlug` with `LoaderPlug` (#1753)
This PR is continuation work of [this
PR](https://github.com/tuist/server/pull/1695) that aims to:
- Consistently use the `selected_` naming convention when assigning the
resources that are targeted by URL and body params.
- Ensure the loading is 1 query (e.g. if we load a run, we load its
project and account in the same query)
- Adds in-memory caching to minimise DB throughput.

This PR gets rid of `EnsureAccountPresencePlug` and moves its logic to
`LoaderPlug`. The only function in it is a function that loads the
account if there's an account handle in the params of `conn`.