Remove crontab worker that's no longer available (#1764)
I removed the `Tuist.CommandEvents.UpdateCacheEventCountWorker` worker
[here](https://github.com/tuist/server/pull/1752/files#diff-89190299739c71927bec9a984eefaf05cf557fcc76332c699f0559212969ae16)
and forgot to delete it from the `crontab` list. For some reason this
went through in `dev` and `test` but launching the app causes it to blow
up. Most likely because Oban is more strict about the presence of the
module in those cases.