Replace Chatwoot with Plain (#1775)
Plain has better insights and integrations that can help us make support
more manageable. This PR replaces the chat box of our marketing site and
added it to the dashboard (we didn't have Chatwoot previously in it).

> [!NOTE]
> I've taken the opportunity to consolidate all the emails into
contact@tuist.dev. We are limited by the amount of emails that we can
use at Plain, which I think it's fine considering our scale. We don't
need an email per department because we don't have those 😬