Increase the Finch pool size for S3 (#1782)
The pool that we have currently setup in production to interact with the
production S3 bucket is limited to 500 connections. This is
underutilising the dedicated hardware in Fly, and potentially causing
some client requests to be
[https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/performance/incide[…]40?selectedTime=2025-04-23T15%3A36%3A00.000Z&activeFilter=date](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/performance/incidents/221/samples/6607b64d83eb6789e61c8ba7-19356908023968840991745422440?selectedTime=2025-04-23T15%3A36%3A00.000Z&activeFilter=date)
because they sit in the server waiting to be assigned to a connection.

I'm bumping the number of connections and keeping the number of pools to
1. I also reached out to the Tigris team to see if they support http2,
which would allow Finch to multiplex multiple requests in the same
connection.