Fix "GenServer DNSCluster terminating" issue (#1798)
At the time when we
[used](https://github.com/tuist/server/commit/bda88ab8deed88b01670f10c1254771fab863f38)
`:nebulex` for caching, I added
[horde](https://hexdocs.pm/nebulex_adapters_horde/Nebulex.Adapters.Horde.html)
to distribute the cache across all the running nodes. This is causing
issues for some enterprise customers where the app fails to launch
because they are not launch the app to be part of a cluster:

```
08:04:55.987 [warning] node not running in distributed mode. Ensure the following exports are set in your rel/env.sh.eex file:

    #!/bin/sh

    export RELEASE_DISTRIBUTION=name
    export RELEASE_NODE="myapp@fully-qualified-host-or-ip"

08:04:55.988 [error] GenServer DNSCluster terminating
```

Since we are not using Horde anymore, let's remove it from the project.