Fix irreversible migrations (#1805)
When removing columns, we _have_ to specify the type of the column or
add explicit `up` and `down` methods. Otherwise, the rollback doesn't
work.

We should run `mix ecto.rollback` for _every_ new migration to make sure
it's reversible. Otherwise, this leads to issues for our enterprise
customers when they need to rollback to some previous version for
whatever reason.