Handle invalid UUIDv7 tokens (#1801)
I noticed some
[errors](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/714/samples/6607b64d83eb6789e61c8ba7-489575373660731413217458491003)
in AppSignal around 9:00 from today because we try to look up an account
token using an ID that's an invalid UUIDv7 string.

I first thought we had introduced a regression somewhere, but there
wasn't any deployment around that time, so it's most likely a client
that's doing something wrong passing the token, or some malicious actor.