Handle missing preview archives (#1803)
The endpoint to download a preview archive assumes that the presence of
a preview translates to the presence of its archive, leading to [server
errors](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/843)
when it does not.

I think we broke that assumption with the migration from Tigris to R2,
as part of which some artifacts couldn't be migrated (which I think is
fine given our current terms of service, something that we are working
on improving), and it might be the case down the road when we establish
a clean up policy.

This PR adjusts that route to handle the scenario more gracefully and
return a 404 in those cases.