Don't spawn `DNSCluster` for on-premise environments (#1812)
Related https://github.com/tuist/server/pull/1807

Tuist continues to abort at startup for on-premise customers because we
still had some code in `application.ex` to start a `DNSCluster` as a
child of the application. For the same reasons [described in this
PR](https://github.com/tuist/server/pull/1807), this is not something we
want to do for on-premise customers.