XCFramework: support signature. (#7411)
* WIP: DO NOT MERGE.

Point to the versions of XcodeGraph and XcodeProject that support this PR
This commit should be removed before merge, it's only until the other PRs are merged

* Support XCFramework expected signature.

When an XCFramework is signed, the signature appears as part of the file reference.
This commit supports the signature by adding two things:
1. Adding the original signature as part of the target dependency object. This
will allow verifying that the signature as extracted from the XCFramework is
the same as expected, and fail at generation time if not.
2. Adding the expectedSignature to the final generated project.

* Fixtures: add signed XCFrameworks fixtures.

Add two fixtures:
1. App with several XCFrameworks dependencies with different signature.
2. App where there's a dependency of an XCFramework with specific signature,
but the actual signature is different.

* TuistCoreTests: Add tests for XCFranework signatures on generation.

* Update dependencies

* Force CI run

* Unify groups

---------

Co-authored-by: Pedro <pedro@pepicrft.me>