Include the project id in the batch key (#1816)
While debugging the failing tests [in this
PR](https://github.com/tuist/cli-ext/pull/382), I realized they fail
because the cache action items are not persisted in the database. It
turns out that [the
batcher](https://appsignal.com/tuist-cloud/sites/66967a1394809a7074e5bc90/exceptions/incidents/177)
fails to run `Repo.insert_all` because some events don't satisfy a
foreign key that ensures the project it refers to is present.

The issue happens because some of our tests in our test suite assume
synchronous insertion of cache action items, leading to the deletion of
the project before the event is processed. This race condition is
unlikely to happen in production (a project is deleted while events are
being processed). Still, I decided to make some changes to handle the
scenario more gracefully. With this PR, all the events that belong to
the same project are processed together. That way, if the processing of
the events of a particular project fails, the failure doesn't prevent
the insertion of other projects' events.

> [!NOTE]
> I'll review the acceptance tests in tuist/cli-ext and remove the
assumption that cache action items are inserted synchronously.