Fix deployments (#1818)
Deployments
[broke](https://github.com/tuist/server/actions/runs/14865674930/job/41741998666)
because we updated the Tuist CLI to the most recent Swift version (which
comes with Xcode 16.3), and it came with some breaking changes in
trailing commas that cause the compilation of acceptance tests to fail.

This PR ensures we are using the same Xcode version in the deployment
pipeline, Xcode 16.3.