Fix the DB configuration (#1819)
I'm still investigating the issue with our DB
([thread](https://github.com/elixir-ecto/db_connection/issues/324)) with
Fly.io. My hunch is that connections are dropped by the network infra
between the Fly and the Supabase instances, and when we get a burst of
queries, which force the connection pool to establish new connections,
some queries end up inevitably dropped from the queue. Note this new
graph that I added where you can see how connections are established
here and there. The disconnections don't show up, but I believe it's a
bug in the connection pool, because they show up in the logs:

```
2025-05-05 18:03:39.979	
16:03:39.977 [error] Postgrex.Protocol (#PID<0.4690.0>) disconnected: ** (DBConnection.ConnectionError) ssl recv (idle): timeout
2025-05-05 18:03:39.979	
16:03:39.977 [error] Postgrex.Protocol (#PID<0.4698.0>) disconnected: ** (DBConnection.ConnectionError) ssl recv (idle): timeout
2025-05-05 18:03:39.979	
16:03:39.977 [error] Postgrex.Protocol (#PID<0.4678.0>) disconnected: ** (DBConnection.ConnectionError) ssl recv (idle): timeout
2025-05-05 18:03:39.979	
16:03:39.977 [error] Postgrex.Protocol (#PID<0.4684.0>) disconnected: ** (DBConnection.ConnectionError) ssl recv (idle): timeout
```

<img width="805" alt="image"
src="https://github.com/user-attachments/assets/75d2f92d-fea6-4ca6-8513-c4b92a132be7"
/>

As part of the investigation I noticed our `parameters` to keep the
connections alive was in the wrong place, so I moved it to the right
place and shortened the times. It's possible that the network
infrastructure is dropping the connections in periods of no activity,
and this is not something we can afford.