Handle nonexistent projects in CleanProjectWorker (#1822)
We have an acceptance tests that ensures cleaning works. This tests
started failing since we made the action asynchronous since on the
client we send a "delete project" right after the completion of the
clean call, leading to a race condition where by the time the job is
picked up by Oban, the project doesn't exist.

This race condition is very unlikely to reproduce in production, but I
added some logic to handle it gracefully anyways.