Redirect from the project overview page if Timescale is not available (#1824)
Timescale is not a strict requirement. This is something I think we
should revisit. In particular:

- Do we want to make it a string requirement?
- Do we want to continue using Timescale

Until we can figure this out with input from existing on-premis
customers, one of them requested if we can redirect to one of the runs
pages instead of overview, which can't be opened if they don't have
Timescale. This PR does that.

As an alternative I considered coming up with some "Timescale not
available" kind of empty state, but I wonder if we should invest in that
direction, or rather wait until we make a decision on what our stance is
regarding storage of analytical data. Thoughts?