Fix displaying preview images (#1866)
Since the JS hook does not load instantly, the fallback image is by
default set as `src` and then switched to the actual image in the hook.
And then back again to the fallback image if the image actually does not
exist.

I'm not too much of a fan of this jumping back and forth, but I don't
have a better solution.

We can (and probably should) track in the db whether a preview icon
exists, so we can show the fallback image eagerly based on the icon's
existence. However, we still need to handle a scenario when loading an
image from the storage fails.