refactor: update sorting of projects and improve query (#1923)
I still want to do pagination, but...
The old page had a few issues:
1. Sorting wasn't working the way I wanted to. It's now all interacted
projects, THEN all projects that haven't been interacted with.
2. It was super inefficient. We'd query CommandEvents for every projects
once for sorting, and then again for the UI. That's 2N queries to
CommandEvents. Oops. This adds a virtual field and does sorting and
querying for all projects in one query. Efficiency.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>