Noora Web announcement blog post (#1937)
Ok ok, I realised I'm not great at writing and after 10 iterations I
still hate it. Fire away all the comments 😂
Considered putting an example of usage but not sure it fits the vibe of
the announcement. Should be easy to add though, so if you think it
should have one, let's go.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>