Fix the worker that updates the monthly cache usage (#1969)
A customer reported that their usage was not being updated. Then I
checked the database and I noticed that many records were not updated.

It turns out that paginating while spawning child jobs that mutate the
rows being paginated is not a good idea. I adjusted the implementation
to create all the jobs in-memory and then insert them all in one shot.