Replace smaller objects with only hiding them from the chart (#1979)
Addresses the feedback from https://github.com/tuist/tuist/issues/7685:
> Most of the time, both in the whole bundle overview and in the
framework breakdown, there is a Smaller object entry that takes most of
the size but there is no way to drill down in what's inside it

The way we solve it is by:
- hiding from the chart actually smaller objects (i.e., we need to order
them before cutting them)
- still showing _all_ objects in the chart table and the file breakdown.
We now only skip showing the elements from the chart. This is similar to
how DaisyDisk solves this. What we currently don't have, and DaisyDisk
does, is to allow you to still drill down into the directories that fall
under "Smaller objects". This would require us to dynamically add/remove
elements from the chart. Instead, I made elements in the chart table not
interactive. I believe that strikes a good balance for now.