Fix multiple previews with same parameters (#1980)
Resolves:
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/924/samples/timestamp/2025-06-19T09:22:00Z

When doing the backfill, each `app_build` was created with its own
`preview`, regardless of if they would fall into the same group. To fix
that, we're limiting the query to one. This should no longer be a
problem for previews shared from new commit SHAs, so it's more of a
temporary fix.