Introduce standardised controller error handling (#1298)
Welcome to the first bigger change I'd like to propose ⭐ 

## Context

Error handling is necessary everywhere and, in most cases, pretty
similar. A general function sequence for controller actions is to fetch
a resource, authorize a user, and do something to the resource. At any
point in this sequence, the call can fail: resource not found, not
allowed to modify it, modification failed. All of these options must be
handled and return a correct, error-specific response to the user.

## Errors in Elixir

Elixir generally handles errors as values. Most functions in the
standard library return `:error` or `{:error, reason}` in case of
failure. Exceptions are very rarely used - in reality, the convention
that has materialised across the community is that exceptions should
only be raised *when the situation is already unrecoverable* and it's
safer to restart the process. You can use `Repo.one!/2` when you are
already sure through other means that the passed identifier exists; if
it still crashes and raises, there's no handling that fixes the issue
(e.g. database connection failing).
In addition, Elixir does not have early returns. This means that calling
`raise` in the middle of a function basically is the last escape hatch
saying "we encountered something that makes the code below useless".

## Proposal: Domain logic

Let's go back to the generic function sequence from above:
1. Fetch a resource. Generally returns either the resource itself, or
`nil`.
2. Authorize a user. Currently, our `Tuist.Authorization` module returns
`true`/`false` for the custom `can/4` implementation, and
`:ok`/`{:error, :unauthorized}` for the policies using LetMe.
3. Modify a resource. In most cases, this will return `{:ok, resource}`,
or `{:error, reason}`, where many times, the reason is an Ecto changeset
with `valid?` set to `false` and errors inside it.

Interestingly, all of these have different return types.  
Let's look at the current implementation of the account controller that
was recently implemented:

```
   with %Account{} = account <- Accounts.get_account_by_handle(handle),
         true <- can(conn.assigns.current_user, :update, account, :organization),
         # The field is called `handle` in the public API, but `name` in our domain.
         {handle, params} = Map.pop(params, :handle),
         params = Map.put(params, :name, handle),
         {:ok, account} <- Accounts.update_account(account, params) do
      conn
      |> put_status(:ok)
      |> json(%{id: account.id, handle: account.name})
    else
      {:error, _changeset} ->
        conn |> put_status(:bad_request) |> json(%{message: "Error updating account."})

      false ->
        conn
        |> put_status(:forbidden)
        |> json(%{message: "You don't have permission to update the #{handle} account."})

      nil ->
        conn
        |> put_status(:not_found)
        |> json(%{message: "An account with this handle was not found."})
    end
 ```

`with` is Elixir's way of doing early returns. You can chain the happy path, and everything that does not match gets matched in the `else` clause, or, alternatively, returned to the function caller. The fact that all of our functions have different return types, though, means a very large `else` clause.  

I therefore propose to, over time, change the domain logic as follows:
1. Getters, instead of returning the resource as a single value or `nil`, return `{:ok, resource}`, and `{:error, :not_found}`.
2. Move all authorization logic to LetMe and set its error reason to `:forbidden` instead of `:unauthorized`.

## Proposal: Interface logic

By default, a Phoenix controller action has to return a `%Plug.Conn{}`. All of the error handling, setting of status codes and building of response bodies needs to happen in the controller so that it can return the conn itself.  
This doesn't have to be that way. By wrapping the action and matching on its result, we can also allow controllers to return error tuples.  

Doing so, with the above proposed changes, we can shorten the above action to:

```
with {:ok, account} <- get_account(handle), # returns {:error,
:not_found}
:ok <- can(conn.assigns.current_user, :update, account, :organization),
# returns {:error, :forbidden}
# The field is called `handle` in the public API, but `name` in our
domain.
         {handle, params} = Map.pop(params, :handle),
         params = Map.put(params, :name, handle),
{:ok, account} <- Accounts.update_account(account, params) do # returns
{:error, changeset}
      conn
      |> put_status(:ok)
      |> json(%{id: account.id, handle: account.name})
    end
 ```

The nature of `with` means that if any of our cases do not match, the
plug calling our action will get one of the following:
- `conn`: our happy path, a conn with an updated account
- `{:error, :not_found}` or `{:error, :not_found, "account"}`: our "we
don't know this" path. I've added a third item to the tuple to customize
the error message.
- `{:error, forbidden}`: you're not allowed to do this
- `{:error, changeset}`: your input is wrong

We can handle all of these generically.  
The most interesting is probably the changeset option where we can
extend the errors we send to the CLI. In the current code draft, I've
set it up to return

```
{
  "message": "There was an error handling your request.",
  "fields": {
    "name": ["Cannot contain a dot."]
  }
}
```

The `message` property remains. This means that this change is
*completely backwards-compatible* and will show the same error as
previously. In addition, we get more information about each field
itself. This allows a CLI return value like

```
There was an error handling your request.
  - Name: Cannot contain a dot.
  - Other field: More information...
```

Let's look at the first case clause of the controller implementation:

```
case apply(__MODULE__, action_name(conn), [conn, conn.params]) do
  %Conn{} = conn ->
    conn
  ...
end
```

If the controller action returns a `%Plug.Conn{}`, we don't touch it and
return it as-is. This, again, means that this change is *completely
backwards-compatible*. All tests pass, since all of our controllers
already return the correct, error-handled response. This *also* means
that any specific, non-generic error handling can still stay in the
controllers. In the above example, having the generic error cases
handled does not stop us from doing something like

```
    with {:ok, account} <- get_account(handle), # returns {:error, :not_found}
         :ok <- can(conn.assigns.current_user, :update, account, :organization), # returns {:error, :forbidden}
         # The field is called `handle` in the public API, but `name` in our domain.
         {handle, params} = Map.pop(params, :handle),
         params = Map.put(params, :name, handle),
         :ok <- some_where_specific_weird_function(), 
         {:ok, account} <- Accounts.update_account(account, params) do # returns {:error, changeset} 
      conn
      |> put_status(:ok)
      |> json(%{id: account.id, handle: account.name})
    else
      {:error, :weird_error} ->
        conn |> put_status(418) |> json(%{message: "I'm a teapot"})
      {:error, other} -> {:error, other} # This will be handled by the controller logic
    end
```

## Implementation

Because this is completely backwards-compatible and doesn't touch any
existing code paths, this proposal does not come with a large refactor.
Actually, it can just be merged as-is without touching anything. This
means that, should we decide to implement a solution like this, can
slowly move over. It can be a "whenever we touch a controller that uses
old error handling, we will also quickly switch it over", or a
"Christoph will put it on his to-do list to move over three actions per
week" or anything in-between.

Excited to read your comments :)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>