Add branch dropdown for bundles (#1985)
Adding a new `Branch` dropdown for the `Bundle size` card: 

![image](https://github.com/user-attachments/assets/bbe2a296-250a-4fa5-9f17-c441eb07000e)

This is useful for two reasons:
- We're explicit about that the by default we show just the bundles for
the `main` branch – instead of doing that implicitly, causing confusion
- If teams upload bundles from other branches, they can still have a
graph across _all_ branches

Ideally, one could choose any _specific_ branch in the dropdown –
however, for that, we'd need a dropdown component with a search which we
have not implemented, yet, in Noora.

Additionally, when there are no bundles in the `main` branch, we hide
the dropdown and default to `Any`. This is very common when initially
setting up the bundle analysis – we want to make a good impression here
and not showing the graph until one merges to `main` is confusing. Based
on feedback at https://github.com/tuist/tuist/issues/7685.

cc'ing @asmitbm if he'd change anything in this flow.