Implement authorization using OAuth 2.0 (#2006)
This PR implements the authorization part of the OAuth 2.0 spec using
[Boruta](https://hexdocs.pm/boruta/readme.html).

The PR also includes all the migrations required for OpenID and the rest
of the OAuth spec like introspection of tokens. We don't have usage for
these, so I'd rather keep the surface smaller for the part of the spec
that we actively use instead of shipping a part of the spec that we
won't be using and thus it could easily get broken. Adding support, once
the need arises, should be straightforward with Boruta generators.

The _very much WIP_ PR that uses the new authorization endpoint from the
iOS app is available here: https://github.com/tuist/tuist/pull/7732

Note that since we're renaming the table for guardian tokens, we should
only merge this over the weekend during low-usage hours. The migration
should be fast, so I don't expect a lot of tokens being generated while
it's running, but it's still safer to run these migrations over the
weekend (plus we're not in a hurry to merge this)

---------

Co-authored-by: Pedro Piñera Buendía <663605+pepicrft@users.noreply.github.com>