Implement Sign in with Apple (#2017)
This PR adds sign in with Apple – both from the browser and from the
app.

The biggest pain point of the integration is that one needs to test
everything from a staging environment – we can't run the sign in from a
localhost because #apple.

For the browser login, we're using ueberauth, this guide was great:
https://github.com/ueberauth/ueberauth_apple/blob/main/guides/getting-started.md#generating-the-client-secret

The authentication from the mobile was a bit convoluted (for me), but
this graph helped me a ton to understand what to do:
https://auth0.com/docs/authenticate/identity-providers/social-identity-providers/apple-native

And here's official documentation for the `token` endpoint:
https://developer.apple.com/documentation/signinwithapplerestapi/generate_and_validate_tokens

---------

Co-authored-by: Pedro Piñera Buendía <663605+pepicrft@users.noreply.github.com>
Co-authored-by: Christoph Schmatzler <christoph@tuist.dev>