Add support for Okta sign in our Tuist-hosted server (#2038)
Adds multi-tenant support for Okta. The docs for how to set it up are at
https://github.com/tuist/tuist/pull/7763. The on-premise customers will
need to update their secrets ever-so-slightly – I'll communicate that
with Zillow who are the only ones using Okta in the on-premise setup.

Want to give a big shoutout to this issue as that helped immensely:
https://github.com/appcues/ueberauth_okta/issues/9

Another source of inspiration was how Deel SSO works:
https://help.letsdeel.com/hc/en-gb/articles/20957859203857-How-to-Setup-OpenID-Connect-Integration-with-Okta-for-Single-Sign-On

The log in page was inspired by Supabase:
https://supabase.com/dashboard/sign-in-sso?returnTo=%2Forg